chore: move lib/internal to base-core/internal
