+++ generators/app/support/config.ts
@@ -1,7 +1,7 @@
 import { upperFirst } from 'lodash-es';
 import { applicationTypes, authenticationTypes, databaseTypes } from '../../../lib/jhipster/index.js';
 import { mutateData } from '../../../lib/utils/index.js';
-import { loadDerivedConfig } from '../../../lib/internal/index.js';
+import { loadDerivedConfig } from '../../base-core/internal/index.js';
 import serverCommand from '../../server/command.js';
 
 const { GATEWAY, MONOLITH } = applicationTypes;
